microservices/email-service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pymongo==4.3.3
pydantic==1.10.7
email-validator==2.0.0
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
httpx==0.24.1
python-multipart==0.0.6